+++ tests/queries/clear_blocks.ts
@@ -1,0 +1,3 @@
+import { SqlUpdate } from "../../src/forms-sql/SqlUpdate";
+
+console.log(SqlUpdate.clearFromBlock("asset", ["abjdfg", "1234556"]));